Field update api

+++ src/main/java/com/group12/springboot/hoversprite/field/FieldCreationRequest.java
@@ -16,8 +16,8 @@ import lombok.Setter;
 public class FieldCreationRequest {
     private Long farmerId;
     private String name;
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
     private String address;
     private CropType cropType;
     private double farmlandArea;
